refactor(orchestration): rename concurrent-task-resolver to concurrent-dep-resolver (refs #3189)
The module under .agents/scripts/lib/orchestration/ exports a single
function, autoSerializeOverlaps, which is a generic graph dep-resolver
that works on any node with focusAreas + dependsOn — not specifically
"tasks". Under the 3-tier hierarchy (Epic -> Feature -> Story) there
is no type::task ticket layer anyway, so the old name was both
inaccurate (it's a dep resolver, not a task resolver) and a stale
echo of the retired 4-tier model.

Decision: rename rather than delete. Live callers remain
(dependency-analyzer.js re-exports it; dispatch-pipeline-reachability
test imports it directly), so the function still earns its keep — only
the name was wrong.

- git mv .agents/scripts/lib/orchestration/concurrent-task-resolver.js
  -> .agents/scripts/lib/orchestration/concurrent-dep-resolver.js
- git mv tests/lib/orchestration/concurrent-task-resolver.test.js
  -> tests/lib/orchestration/concurrent-dep-resolver.test.js
- Update import in dependency-analyzer.js re-export.
- Update import in dispatch-pipeline-reachability.test.js.
- Update test description prefixes to match new module name.

Baselines under baselines/ still reference the old path; those files
are regenerated by the baseline pipeline and will refresh on the next
scheduled regen (no behavior change).
