+++ .agents/scripts/lib/orchestration/concurrent-dep-resolver.js
